assets: add support for configurable bufsize for FileHandler
The current implementation is using the blocksize as buffer size.
In many use cases, the size is too small and limiting throughput.
This commit adds an new option, buffer_size in the FileOptions, which,
when set, overrides the default value.

Also included is a cargo criterion benchmark, which can be run with

```
cargo bench --bench file_handler
```

The benchmark shows major performance improvements with larger buffers
when serving large files.

Signed-off-by: Janne Pelkonen <jpelkonen@ideanovatech.com>